# src/main/java/io/swagger/api/ClienteApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.19).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Cliente;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2021-04-07T05:48:22.248Z")

@Validated
@Api(value = "cliente", description = "the cliente API")
@RequestMapping(value = "/cliente-api/v1")
public interface ClienteApi {

    @ApiOperation(value = "Alteração de status de cliente por id.", nickname = "alteraStatusPorId", notes = "Esta operação tem por objetivo alterar o status do cliente por id.", response = Cliente.class, tags={ "Manutencao", })
    @ApiResponses(value = { 
        @ApiResponse(code = 202, message = "Sucesso ao alterar o status de um cliente existente.", response = Cliente.class),
        @ApiResponse(code = 400, message = "Alteração de status inválido."),
        @ApiResponse(code = 401, message = "Alteração não autorizada."),
        @ApiResponse(code = 500, message = "Erro interno ao tentar realizar a alteração de status.") })
    @RequestMapping(value = "/cliente/{status}/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<Cliente> alteraStatusPorId(@ApiParam(value = "Status do cliente.",required=true, allowableValues = "\"ativo\", \"inativo\"") @PathVariable("status") String status,@ApiParam(value = "Id do cliente.",required=true) @PathVariable("id") Integer id);


    @ApiOperation(value = "Altera um cliente existente", nickname = "alterarCliente", notes = "Esta operação tem por objetivo alterar um cliente existente.", response = Cliente.class, tags={ "Manutencao", })
    @ApiResponses(value = { 
        @ApiResponse(code = 202, message = "Sucesso ao alterar um cliente existente.", response = Cliente.class),
        @ApiResponse(code = 400, message = "Alteração inválido."),
        @ApiResponse(code = 401, message = "Alteração não autorizada."),
        @ApiResponse(code = 404, message = "Nenhum cliente encontrado para Alteração."),
        @ApiResponse(code = 500, message = "Erro interno ao tentar realizar Alteração.") })
    @RequestMapping(value = "/cliente/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<Cliente> alterarCliente(@ApiParam(value = "Id do cliente.",required=true) @PathVariable("id") Integer id,@ApiParam(value = "" ,required=true )  @Valid @RequestBody Cliente cliente);


    @ApiOperation(value = "Cadastro de um novo cliente", nickname = "cadastrarNovo", notes = "Esta operação tem por objetivo cadastrar um novo cliente", response = Cliente.class, tags={ "Cadastro", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Sucesso ao cadastrar um novo cliente.", response = Cliente.class),
        @ApiResponse(code = 400, message = "Cadastro inválido."),
        @ApiResponse(code = 401, message = "Cadastro não autorizada."),
        @ApiResponse(code = 404, message = "Nenhum cliente encontrado para a cadastro."),
        @ApiResponse(code = 500, message = "Erro interno ao tentar realizar a cadastro.") })
    @RequestMapping(value = "/cliente",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<Cliente> cadastrarNovo(@ApiParam(value = "" ,required=true )  @Valid @RequestBody Cliente cliente);


    @ApiOperation(value = "Consulta um cliente existente por id.", nickname = "consultarPorId", notes = "Esta consulta tem por objetivo consultar um cliente existente por id.", response = Cliente.class, tags={ "Consulta", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Sucesso ao consultar um cliente existente por id.", response = Cliente.class),
        @ApiResponse(code = 400, message = "Consulta inválido."),
        @ApiResponse(code = 401, message = "Consulta não autorizada."),
        @ApiResponse(code = 404, message = "Nenhum cliente encontrado para consulta."),
        @ApiResponse(code = 500, message = "Erro interno ao tentar realizar consulta.") })
    @RequestMapping(value = "/cliente/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.GET)
    ResponseEntity<Cliente> consultarPorId(@ApiParam(value = "Id do cliente.",required=true) @PathVariable("id") Integer id);


    @ApiOperation(value = "Consulta cliente existente por sobrenome.", nickname = "consultarPorSobrenome", notes = "Esta operação tem por obejtivo consultar o cliente por sobrenome", response = Cliente.class, tags={ "Consulta", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Sucesso ao consultar um cliente existente por sobrenome", response = Cliente.class),
        @ApiResponse(code = 400, message = "Consulta inválido."),
        @ApiResponse(code = 401, message = "Exclusão não autorizada."),
        @ApiResponse(code = 404, message = "Nenhum cliente encontrado para consulta."),
        @ApiResponse(code = 500, message = "Erro interno ao tentar realizar consulta.") })
    @RequestMapping(value = "/cliente/consulta/{sobrenome}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.GET)
    ResponseEntity<Cliente> consultarPorSobrenome(@ApiParam(value = "sobrenome do cliente.",required=true) @PathVariable("sobrenome") String sobrenome);


    @ApiOperation(value = "Exclui um cliente existente", nickname = "excluir", notes = "Esta operação tem por objetivo exlcuir um cliente existente.", tags={ "Exclusao", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "Sucesso ao excluir um cliente existente."),
        @ApiResponse(code = 400, message = "Exclusão inválido."),
        @ApiResponse(code = 401, message = "Exclusão não autorizada."),
        @ApiResponse(code = 404, message = "Nenhum cliente encontrado para exclusão."),
        @ApiResponse(code = 500, message = "Erro interno ao tentar realizar exclusão.") })
    @RequestMapping(value = "/cliente/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.DELETE)
    ResponseEntity<Void> excluir(@ApiParam(value = "Id do cliente.",required=true) @PathVariable("id") Integer id);

}