# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'Cliente API'
  description: '|
  Esta api tem por objetivo cadastrar, consultar, alterar e excluir clientes.
  [Saiba mais](https://github.com/ValneiPereira)
  '
consumes:
    - application/json
produces:
    - application/json
    
tags: 
  - name: 'Consulta'
  - name: 'Cadastro'
  - name: 'Manutencao'
  - name: 'Exclusao'
  - name: 'API'
  
# localhost:8080/cliente-api/v1
host: localhost:8080
basePath: /cliente-api/v1
schemes: 
  - http
  - https
paths:
  # localhost:8080/cliente-api/v1/clientes
  /clientes:
    get:
      tags:
        - Consulta
      summary : "Consulta todos os clientes."
      operationId: "consultarTodos"
      description: "Esta operação tem por objetivo consultar todos os clientes"
      responses:
        200:
          description: "Sucesso ao consultar todos os clientes."
          schema:
            $ref: "#/definitions/Clientes"
        400:
          description: "Consulta inválida."
        401:
          description: "Consulta não autorizada."
        404:
          description: "Nenhum cliente encontrado para a consulta."
        500:
          description: "Erro interno ao tentar realizar a consulta."
          
  /cliente:
    post:
      tags:
        - Cadastro
      summary: "Cadastro de um novo cliente"
      operationId: "cadastrarNovo"
      description: "Esta operação tem por objetivo cadastrar um novo cliente"
      parameters:
        - in: "body"
          name: "cliente"
          required: true
          schema:
            $ref: "#/definitions/Cliente"
      responses:
        201:
          description: "Sucesso ao cadastrar um novo cliente."
          schema:
            $ref: "#/definitions/Cliente"
          headers:
            Location:
              type: "string"
              description: "URI do cliente cadastrado."
        400:
          description: "Cadastro inválido."
        401:
          description: "Cadastro não autorizada."
        404:
          description: "Nenhum cliente encontrado para a cadastro."
        500:
          description: "Erro interno ao tentar realizar a cadastro."
  
  /cliente/{id}:
    put:
      tags:
        - Manutencao
      summary: "Altera um cliente existente"
      operationId: "alterarCliente"
      description: "Esta operação tem por objetivo alterar um cliente existente."
      parameters:
        - in: "path"
          name: "id"
          type: "integer"
          format: "int32"
          required: true
          description: "Id do cliente."
        - in: "body"
          name: "cliente"
          required: true
          schema:
            $ref: "#/definitions/Cliente"
      responses:
        202:
          description: "Sucesso ao alterar um cliente existente."
          schema:
            $ref: "#/definitions/Cliente"
          headers:
            Location:
              type: "string"
              description: "URI do cliente alterado."
        400:
          description: "Alteração inválido."
        401:
          description: "Alteração não autorizada."
        404:
          description: "Nenhum cliente encontrado para Alteração."
        500:
          description: "Erro interno ao tentar realizar Alteração."
    get:
      tags:
        - Consulta
      summary: "Consulta um cliente existente por id."
      operationId: "consultarPorId"
      description: "Esta consulta tem por objetivo consultar um cliente existente por id."
      parameters:
        - in: "path"
          name: "id"
          type: "integer"
          format: "int32"
          required: true
          description: "Id do cliente."
      responses:
        200:
          description: "Sucesso ao consultar um cliente existente por id."
          schema:
            $ref: "#/definitions/Cliente"
        400:
          description: "Consulta inválido."
        401:
          description: "Consulta não autorizada."
        404:
          description: "Nenhum cliente encontrado para consulta."
        500:
          description: "Erro interno ao tentar realizar consulta."
          
    delete:
      tags:
        - Exclusao
      summary: "Exclui um cliente existente"
      operationId: "excluir"
      description: "Esta operação tem por objetivo exlcuir um cliente existente."
      parameters:
        - in: "path"
          name: "id"
          type: "integer"
          format: "int32"
          required: true
          description: "Id do cliente."
      responses:
        204:
          description: "Sucesso ao excluir um cliente existente."
        400:
          description: "Exclusão inválido."
        401:
          description: "Exclusão não autorizada."
        404:
          description: "Nenhum cliente encontrado para exclusão."
        500:
          description: "Erro interno ao tentar realizar exclusão."
          
  /cliente/consulta/{sobrenome}:
    get:
      tags:
      - Consulta
      summary: "Consulta cliente existente por sobrenome."
      operationId: "consultarPorSobrenome"
      description: "Esta operação tem por obejtivo consultar o cliente por sobrenome"
      parameters:
        - in: "path"
          name: "sobrenome"
          type: "string"
          required: true
          description: "sobrenome do cliente."
      responses:
        200:
          description: "Sucesso ao consultar um cliente existente por sobrenome"
          schema:
            $ref: "#/definitions/Cliente"
        400:
          description: "Consulta inválido."
        401:
          description: "Exclusão não autorizada."
        404:
          description: "Nenhum cliente encontrado para consulta."
        500:
          description: "Erro interno ao tentar realizar consulta."
    
  /cliente/{status}/{id}:
    put:
      tags:
      - Manutencao
      summary: "Alteração de status de cliente por id."
      operationId: "alteraStatusPorId"
      description: "Esta operação tem por objetivo alterar o status do cliente por id."
      parameters:
        - in: "path"
          name: "status"
          type: "string"
          enum: ["ativo", "inativo"]
          required: true
          description: "Status do cliente."
        - in: "path"
          name: "id"
          type: "integer"
          format: "int32"
          required: true
          description: "Id do cliente."
      responses:
        202:
          description: "Sucesso ao alterar o status de um cliente existente."
          schema:
            $ref: "#/definitions/Cliente"
          headers:
            Location:
              type: "string"
              description: "URI do cliente alterado."
        400:
          description: "Alteração de status inválido."
        401:
          description: "Alteração não autorizada."
        500:
          description: "Erro interno ao tentar realizar a alteração de status."
  /readme:
    get:
      tags:
        - API
      summary: "Recupera readme da API."
      operationId: "readme"
      description: "Esta operação retorna o readme da API."
      responses:
        200:
          description: "Sucesso ao recuperar readme da API."
        500:
          description: "Erro interno ao tentar realizar recuperar o readme da API."
          
definitions:
  Clientes:
    type: array
    items:
      $ref: '#/definitions/Cliente'
      
  Cliente:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
        description: "Identificador do cliente."
      nome:
        type: "string"
        description: "Nome do cliente."
      sobrenome:
        type: "string"
        description: "sobre nome do cliente."
      iniciais:
        type: "string"
        minLength: 2
        maxLength: 7
        description: "Iniciais do  nome completo."
      dataNascimento:
        type: "string"
        format: "date"
        description: "Data de nascimento do cliente."
      dataAtualizacao:
        type: "string"
        format: "date-time"
        description: "Data de atualização dos dados do cliente."
      status:
        type: boolean
        description: "status ativo ou inativo do cliente."
      tipo:
        type: "string"
        enum: ["interno", "externo"]